requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
blinker==1.6.2
click==8.1.3
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pip==23.1.2
setuptools==68.0.0
Werkzeug==2.3.6
wheel==0.40.0
zipp==3.15.0
python-dotenv~=1.0.0
